requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain_google_genai
google-generativeai
-e .